nt36xxx: Add `pm_mutex` to struct nvt_ts_data
At the guard of TOUCHSCREEN_LAVENDER, pm_mutex is tried to be accessed
using an arrow operator from struct nvt_ts_data which is referenced
by the pointer in question while locking and unlocking the mutex.

But the member was found missing by the compiler which caused errors.
Hence, add the rightful member at the struct to fix the following compilation
errors when compiling using lavender-perf_defconfig:

../drivers/input/touchscreen/nt36xxx/nt36xxx.c:1086:18: error: no member named 'pm_mutex' in 'struct nvt_ts_data'
        mutex_lock(&ts->pm_mutex);
                    ~~  ^
../drivers/input/touchscreen/nt36xxx/nt36xxx.c:1088:20: error: no member named 'pm_mutex' in 'struct nvt_ts_data'
        mutex_unlock(&ts->pm_mutex);
                      ~~  ^
../drivers/input/touchscreen/nt36xxx/nt36xxx.c:1105:20: error: no member named 'pm_mutex' in 'struct nvt_ts_data'
                        mutex_lock(&ts->pm_mutex);
                                    ~~  ^
../drivers/input/touchscreen/nt36xxx/nt36xxx.c:1109:22: error: no member named 'pm_mutex' in 'struct nvt_ts_data'
                        mutex_unlock(&ts->pm_mutex);
                                      ~~  ^

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>