power: supply: Conditionally compile logging
We have POWER_SUPPLY_DEBUG config option available at the
Kconfig of drivers/power.
Use it with an if-else conditional check to compile logging
only when CONFIG_POWER_SUPPLY_DEBUG is enabled. These spew quite
the useless logs which we won't ever read and we don't them to be
available on production builds.