[FIXUP] 074e6e411a ("usb: dwc3: gadget: Do link recovery for SS and SSP")
commit d0550cd20e52558ecf6847a0f96ebd5d944c17e4 upstream.

The controller always supports link recovery for device in SS and SSP.
Remove the speed limit check. Also, when the device is in RESUME or
RESET state, it means the controller received the resume/reset request.
The driver must send the link recovery to acknowledge the request. They
are valid states for the driver to send link recovery.

Fixes: 72246da ("usb: Introduce DesignWare USB3 DRD Driver")
Fixes: ee5cd41 ("usb: dwc3: Update speed checks for SuperSpeedPlus")
Signed-off-by: Thinh Nguyen <thinhn@synopsys.com>
Signed-off-by: Felipe Balbi <balbi@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
[ Tashar02: This was broken with commit 1234119 ("ACK: Merge tag `v4.19.237` of android-4.19-stable") ]
Change-Id: I5831a97c1e6dbfcd8377dcd5bc33ddee731246f0
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>